adds initial styles to our app
