+++ frontend/src/components/Hero/style.css
@@ -1,0 +1,14 @@
+.container-hero {
+  display: flex;
+  flex-direction: column;
+  margin: 10px;
+}
+
+.container-hero img {
+  opacity: 1;
+}
+
+.container-hero img:hover {
+  opacity: 0.5;
+  cursor: pointer;
+}
