+++ frontend/src/components/HeroesList/HeroesList.js
@@ -1,13 +1,15 @@
 import React, { useState } from 'react'
-import Hero from './Hero'
+import Hero from '../Hero/Hero'
+
+import './style.css'
 
 const HeroesList = ({ heroes }) => {
 
 	const [currentPage, setCurrentPage] = useState(1)
 
-	const heroesPerPage = 1
+	const heroesPerPage = 5
 	const lastHeroIndex = currentPage * heroesPerPage;
-  const firstHeroIndex = lastHeroIndex - heroesPerPage;
+	const firstHeroIndex = lastHeroIndex - heroesPerPage;
 	const currentHeroes = heroes.slice(firstHeroIndex, lastHeroIndex)
 	const lastPage = Math.ceil(heroes.length / heroesPerPage) + 1
 
@@ -21,18 +23,20 @@ const HeroesList = ({ heroes }) => {
 
 	return (
 		<div>
-			{heroes.length === 0 ? (
-				<h2>There are no heroes to show</h2>
-			) :
-			currentHeroes.map(hero => (
-					<Hero key={hero.id} hero={hero} />
-				)
-			)
-			}
+			<div className="container" >
+				{heroes.length === 0 ? (
+					<h2>There are no heroes to show</h2>
+				) :
+					currentHeroes.map(hero => (
+						<Hero key={hero.id} hero={hero} />
+					)
+					)
+				}
+			</div>
 			{currentPage === 1 ? (<div></div>) : (<button onClick={handleBack}> Back </button>)}
-			{currentPage +1 !== lastPage ? (<button onClick={handleNext}> Next </button>) : (<div></div>)}
-			
-			
+			{currentPage + 1 !== lastPage ? (<button onClick={handleNext}> Next </button>) : (<div></div>)}
+
+
 		</div>
 	)
