+++ frontend/src/components/HeroesList/style.css
@@ -1,0 +1,6 @@
+.container {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+}
