+++ frontend/src/components/SearchBar/SearchBar.js
@@ -1,8 +1,10 @@
 import React from 'react'
+import './style.css'
 
 const SearchBar = ({ superheroName, setSuperheroName}) => {
 	return (
 		<input
+			className='header-input'
 			placeholder="search superhero"
 			value={superheroName}
 			onChange={e => setSuperheroName(e.target.value)} />
