+++ frontend/src/layout/Home/style.css
@@ -1,0 +1,6 @@
+.header {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+}
